style: rename theme settings to preferences
