+++ app/src/main/java/com/github/se/assocify/ui/screens/profile/ProfileGraph.kt
@@ -8,9 +8,9 @@ import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.screens.profile.members.profileMembersGraph
 import com.github.se.assocify.ui.screens.profile.notifications.profileNotificationsGraph
+import com.github.se.assocify.ui.screens.profile.preferences.profilePreferencesGraph
 import com.github.se.assocify.ui.screens.profile.roles.profileRolesGraph
 import com.github.se.assocify.ui.screens.profile.securityPrivacy.profileSecurityPrivacyGraph
-import com.github.se.assocify.ui.screens.profile.theme.profileThemeGraph
 
 fun NavGraphBuilder.profileGraph(
     navigationActions: NavigationActions,
@@ -25,7 +25,7 @@ fun NavGraphBuilder.profileGraph(
 
   profileNotificationsGraph(navigationActions)
   profileSecurityPrivacyGraph(navigationActions)
-  profileThemeGraph(navigationActions)
+  profilePreferencesGraph(navigationActions)
   profileRolesGraph(navigationActions)
   profileMembersGraph(navigationActions)
 }
