--- app/src/main/java/com/github/se/assocify/ui/screens/profile/theme/ProfileThemeGraph.kt
@@ -1,12 +1,0 @@
-package com.github.se.assocify.ui.screens.profile.theme
-
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.github.se.assocify.navigation.Destination
-import com.github.se.assocify.navigation.NavigationActions
-
-fun NavGraphBuilder.profileThemeGraph(navigationActions: NavigationActions) {
-  composable(route = Destination.ProfileTheme.route) {
-    ProfileThemeScreen(navActions = navigationActions)
-  }
-}
